Move user model into /models.
