+++ lib/services/user/firebase_user_database.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:film_freund/services/date_time.dart/date_time_service.dart';
 import 'package:film_freund/services/service_locator.dart';
-import 'package:film_freund/services/user/user.dart';
+import 'package:film_freund/services/user/models/user.dart';
 
 import 'i_user_database.dart';
