+++ lib/services/user/i_user_database.dart
@@ -1,4 +1,4 @@
-import 'package:film_freund/services/user/user.dart';
+import 'package:film_freund/services/user/models/user.dart';
 
 /// A service to perform operations on the [User] database
 abstract class IUserDatabase {
